requirements.txt
aiojobs==1.0.0
aioredis==1.3.1
alembic==1.7.7
alembic-utils==0.7.7
amqp==5.1.1
anyio==3.6.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.6
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colorlog==6.6.0
contextlib2==21.6.0
coverage==6.3.3
defusedxml==0.7.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
Faker==13.11.1
fakeredis==1.7.5
fastapi==0.78.0
fastapi-contrib==0.2.11
fastapi-plugins==0.9.0
fastapi-route-logger-middleware==0.2.3
flupy==1.1.9
future==0.18.2
googleapis-common-protos==1.56.1
graypy==2.1.0
greenlet==1.1.2
grpcio==1.46.1
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
isodate==0.6.1
kombu==5.2.4
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
opentelemetry-api==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-exporter-jaeger-proto-grpc==1.11.1
opentelemetry-exporter-jaeger-thrift==1.11.1
opentelemetry-instrumentation==0.30b1
opentelemetry-instrumentation-asgi==0.30b1
opentelemetry-instrumentation-fastapi==0.30b1
opentelemetry-instrumentation-httpx==0.30b1
opentelemetry-instrumentation-requests==0.30b1
opentelemetry-instrumentation-sqlalchemy==0.30b1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
opentelemetry-util-http==0.30b1
packaging==21.3
parse==1.19.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-cov==3.0.0
pytest-mock==3.7.0
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.2
pytz==2022.1
redis==4.3.1
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
sentry-sdk==1.5.12
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
sqlalchemy-datatables==2.0.1
SQLAlchemy-serializer==1.4.1
starlette==0.19.1
tenacity==8.0.1
text-unidecode==1.3
threadloop==1.0.2
thrift==0.16.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xmltodict==0.13.0
zeep==4.1.0
zipp==3.8.0
